Move RX/TX Lua common into another unit
