+++ src/lib/LUA/tx_devLUA.cpp
@@ -1,15 +1,9 @@
 #ifdef TARGET_TX
 
-#include "common.h"
-#include "device.h"
-
+#include "rxtx_devLua.h"
 #include "CRSF.h"
-#include "POWERMGNT.h"
-#include "config.h"
 #include "logging.h"
-#include "lua.h"
 #include "OTA.h"
-#include "hwTimer.h"
 #include "FHSS.h"
 
 static char version_domain[20+1+6+1];
@@ -54,7 +48,7 @@ static struct luaItem_selection luaDynamicPower = {
     {"Dynamic", CRSF_TEXT_SELECTION},
     0, // value
     "Off;Dyn;AUX9;AUX10;AUX11;AUX12",
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 #if defined(GPIO_PIN_FAN_EN)
@@ -62,14 +56,14 @@ static struct luaItem_selection luaFanThreshold = {
     {"Fan Thresh", CRSF_TEXT_SELECTION},
     0, // value
     "10mW;25mW;50mW;100mW;250mW;500mW;1000mW;2000mW;Never",
-    emptySpace // units embedded so it won't display "NevermW"
+    STR_EMPTYSPACE // units embedded so it won't display "NevermW"
 };
 #endif
 
 #if defined(Regulatory_Domain_EU_CE_2400)
 static struct luaItem_string luaCELimit = {
     {"100mW CE LIMIT", CRSF_INFO},
-    emptySpace
+    STR_EMPTYSPACE
 };
 #endif
 
@@ -79,7 +73,7 @@ static struct luaItem_selection luaSwitch = {
     {"Switch Mode", CRSF_TEXT_SELECTION},
     0, // value
     switchmodeOpts4ch,
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_selection luaModelMatch = {
@@ -92,12 +86,12 @@ static struct luaItem_selection luaModelMatch = {
 static struct luaItem_command luaBind = {
     {"Bind", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_string luaInfo = {
     {"Bad/Good", (crsf_value_type_e)(CRSF_INFO | CRSF_FIELD_ELRS_HIDDEN)},
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_string luaELRSversion = {
@@ -114,26 +108,26 @@ static struct luaItem_folder luaWiFiFolder = {
 static struct luaItem_command luaWebUpdate = {
     {"Enable WiFi", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 #endif
 
 static struct luaItem_command luaRxWebUpdate = {
     {"Enable Rx WiFi", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_command luaTxBackpackUpdate = {
     {"Enable Backpack WiFi", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_command luaVRxBackpackUpdate = {
     {"Enable VRx WiFi", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 //---------------------------- WiFi -----------------------------
 
@@ -141,7 +135,7 @@ static struct luaItem_command luaVRxBackpackUpdate = {
 static struct luaItem_command luaBLEJoystick = {
     {"BLE Joystick", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 #endif
 
@@ -154,21 +148,21 @@ static struct luaItem_selection luaVtxBand = {
     {"Band", CRSF_TEXT_SELECTION},
     0, // value
     "Off;A;B;E;F;R;L",
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_selection luaVtxChannel = {
     {"Channel", CRSF_TEXT_SELECTION},
     0, // value
     "1;2;3;4;5;6;7;8",
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_selection luaVtxPwr = {
     {"Pwr Lvl", CRSF_TEXT_SELECTION},
     0, // value
     "-;1;2;3;4;5;6;7;8",
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_selection luaVtxPit = {
@@ -179,13 +173,13 @@ static struct luaItem_selection luaVtxPit = {
     ";AUX5" LUASYM_ARROW_UP ";AUX5" LUASYM_ARROW_DN ";AUX6" LUASYM_ARROW_UP ";AUX6" LUASYM_ARROW_DN
     ";AUX7" LUASYM_ARROW_UP ";AUX7" LUASYM_ARROW_DN ";AUX8" LUASYM_ARROW_UP ";AUX8" LUASYM_ARROW_DN
     ";AUX9" LUASYM_ARROW_UP ";AUX9" LUASYM_ARROW_DN ";AUX10" LUASYM_ARROW_UP ";AUX10" LUASYM_ARROW_DN,
-    emptySpace
+    STR_EMPTYSPACE
 };
 
 static struct luaItem_command luaVtxSend = {
     {"Send VTx", CRSF_COMMAND},
     lcsIdle, // step
-    emptySpace
+    STR_EMPTYSPACE
 };
 //----------------------------VTX ADMINISTRATOR------------------
 
@@ -194,7 +188,7 @@ struct luaItem_selection luaBluetoothTelem = {
     {"BT Telemetry", CRSF_TEXT_SELECTION},
     0, // value
     "Off;On",
-    emptySpace
+    STR_EMPTYSPACE
 };
 #endif
 
@@ -207,19 +201,19 @@ static struct luaItem_selection luaDvrAux = {
     {"DVR AUX", CRSF_TEXT_SELECTION},
     0, // value
     "Off;AUX1;!AUX1;AUX2;!AUX2;AUX3;!AUX3;AUX4;!AUX4;AUX5;!AUX5;AUX6;!AUX6;AUX7;!AUX7;AUX8;!AUX8;AUX9;!AUX9;AUX10;!AUX10",
-    emptySpace};
+    STR_EMPTYSPACE};
 
 static struct luaItem_selection luaDvrStartDelay = {
     {"DVR Srt Dly", CRSF_TEXT_SELECTION},
     0, // value
     "0s;5s;15s;30s;45s;1min;2min",
-    emptySpace};
+    STR_EMPTYSPACE};
 
 static struct luaItem_selection luaDvrStopDelay = {
     {"DVR Stp Dly", CRSF_TEXT_SELECTION},
     0, // value
     "0s;5s;15s;30s;45s;1min;2min",
-    emptySpace};
+    STR_EMPTYSPACE};
 
 //---------------------------- BACKPACK ------------------
 
@@ -330,7 +324,7 @@ static void luahandWifiBle(struct luaPropertiesCommon *item, uint8_t arg)
       break;
 
     case lcsCancel:
-      sendLuaCommandResponse(cmd, lcsIdle, emptySpace);
+      sendLuaCommandResponse(cmd, lcsIdle, STR_EMPTYSPACE);
       if (connectionState == targetState)
       {
         rebootTime = millis() + 400;
@@ -378,7 +372,7 @@ static void luahandSimpleSendCmd(struct luaPropertiesCommon *item, uint8_t arg)
   } /* if doExecute */
   else if(arg == lcsCancel || ((millis() - lastLcsPoll)> 2000))
   {
-    sendLuaCommandResponse((struct luaItem_command *)item, lcsIdle, emptySpace);
+    sendLuaCommandResponse((struct luaItem_command *)item, lcsIdle, STR_EMPTYSPACE);
   }
 }
